Shopping list module

+++ src/main/java/com/pro/list_tick/shopping_list/service/implementation/ShoppingListServiceImpl.java
@@ -1,22 +1,19 @@
 package com.pro.list_tick.shopping_list.service.implementation;
 
-import com.pro.list_tick.shared.api.AccountAPI;
 import com.pro.list_tick.shared.current_user.CurrentAccountService;
 import com.pro.list_tick.shopping_list.dto.AccountSharedWithDto;
 import com.pro.list_tick.shopping_list.dto.ItemDTO;
 import com.pro.list_tick.shopping_list.dto.ShoppingListDTO;
 import com.pro.list_tick.shopping_list.dto.ShoppingListInputDTO;
 import com.pro.list_tick.shopping_list.dto.ShoppingListUpdateDTO;
-import com.pro.list_tick.shopping_list.exception.CategoryException;
 import com.pro.list_tick.shopping_list.exception.ShoppingListException;
-import com.pro.list_tick.shopping_list.mapper.ItemMapper;
 import com.pro.list_tick.shopping_list.mapper.ShoppingListMapper;
-import com.pro.list_tick.shopping_list.model.Category;
 import com.pro.list_tick.shopping_list.model.SharedShoppingList;
 import com.pro.list_tick.shopping_list.model.ShoppingList;
-import com.pro.list_tick.shopping_list.repository.SLCategoryRepository;
-import com.pro.list_tick.shopping_list.repository.SharedShoppingListRepository;
 import com.pro.list_tick.shopping_list.repository.ShoppingListRepository;
+import com.pro.list_tick.shopping_list.service.CategoryService;
+import com.pro.list_tick.shopping_list.service.ItemService;
+import com.pro.list_tick.shopping_list.service.SharedShoppingListService;
 import com.pro.list_tick.shopping_list.service.ShoppingListService;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -34,65 +31,59 @@ import java.util.UUID;
 @Slf4j
 public class ShoppingListServiceImpl implements ShoppingListService {
 
-    public static final String LIST_NOT_FOUND = "Shopping list not found: %s";
-    public static final String LIST_CONFLICT = "Shopping list not found: %s, for the user: %s";
+    private final ShoppingListRepository shoppingListRepository;
 
     private final CurrentAccountService currentAccountService;
-    private final AccountAPI accountAPI;
+    private final CategoryService categoryService;
+    private final ItemService itemService;
+    private final SharedShoppingListService sharedShoppingListService;
 
-    private final ShoppingListRepository shoppingListRepository;
-    private final SLCategoryRepository categoryRepository;
-    private final SharedShoppingListRepository sharedShoppingListRepository;
+    public ShoppingList getById(UUID id) {
+        log.debug("Getting the shopping list: {}", id);
+        final var shoppingList = shoppingListRepository.findById(id)
+            .orElseThrow(() -> new ShoppingListException(String.format("Shopping list not found: %s", id)));
 
-    public List<ShoppingListDTO> getAllByAccountId() {
+        final var accountId = currentAccountService.getCurrentAccountId();
+        if (!validateAccess(accountId, shoppingList) &&
+            !validateSharedAccess(accountId, shoppingList)) {
+            throw new ShoppingListException(HttpStatus.FORBIDDEN,
+                String.format("User doesn't have an access to the shopping list: %s", shoppingList.getId()));
+        }
+        return shoppingList;
+    }
+
+    public List<ShoppingListDTO> getAllDTOByAccountId() {
         final var accountId = currentAccountService.getCurrentAccountId();
         log.debug("Getting all shopping lists for the accountId: {}", accountId);
 
         var shoppingLists = shoppingListRepository.findAllByAccountId(accountId);
-        var sharedShoppingLists = sharedShoppingListRepository.findAllByIdAccountId(accountId);
-        List<ShoppingListDTO> dtoList = new ArrayList<>(shoppingLists.stream().map(ShoppingListMapper::toDTO).toList());
-        dtoList.addAll(sharedShoppingLists.stream().map(SharedShoppingList::getShoppingList).map(ShoppingListMapper::toDTO).toList());
+        var sharedShoppingLists = sharedShoppingListService.getAllByAccountId(accountId);
+        List<ShoppingListDTO> dtoList = new ArrayList<>(shoppingLists.stream()
+            .map(ShoppingListMapper::toDTO)
+            .toList());
+        dtoList.addAll(sharedShoppingLists.stream()
+            .map(SharedShoppingList::getShoppingList)
+            .map(ShoppingListMapper::toDTO)
+            .toList());
         return dtoList;
     }
 
-    public ShoppingListDTO getById(UUID id) {
-        log.debug("Getting the shopping list: {}", id);
-        final var shoppingList = shoppingListRepository.findById(id)
-                .orElseThrow(() -> new ShoppingListException(String.format(LIST_NOT_FOUND, id)));
-        final var accountId = currentAccountService.getCurrentAccountId();
-        if (!shoppingList.getAccountId().equals(accountId)) {
-            var errorMessage = String.format(LIST_CONFLICT, shoppingList.getId(), accountId);
-            log.error(errorMessage);
-            throw new ShoppingListException(HttpStatus.CONFLICT, errorMessage);
-        }
-        return ShoppingListMapper.toDTO(shoppingList);
-    }
-
     public List<ItemDTO> getItemsByShoppingListId(UUID id) {
-        log.debug("Getting items by shopping list id: {}", id);
-        final var accountId = currentAccountService.getCurrentAccountId();
-        if (!shoppingListRepository.existsByIdAndAccountId(id, accountId)) {
-            var errorMessage = String.format(LIST_CONFLICT, id, accountId);
-            log.error(errorMessage);
-            throw new ShoppingListException(HttpStatus.CONFLICT, errorMessage);
-        }
-        var shoppingList = shoppingListRepository.findByIdWithItems(id)
-                .orElseThrow(() -> new ShoppingListException(HttpStatus.NOT_FOUND, String.format(LIST_NOT_FOUND, id)));
-        return shoppingList.getItems().stream().map(ItemMapper::toDTO).toList();
+        return itemService.getAllByShoppingListId(id);
     }
 
     @Transactional(transactionManager = "shoppingListTransactionManager")
     public ShoppingListDTO create(ShoppingListInputDTO shoppingListInputDTO) {
         final var accountId = currentAccountService.getCurrentAccountId();
-        log.info("Creating a shopping list for the account id: {}, name: {}",accountId, shoppingListInputDTO.getName());
+        log.info("Creating a shopping list for the account id: {}, name: {}",
+            accountId, shoppingListInputDTO.getName());
 
         var shoppingList = ShoppingListMapper.toModel(shoppingListInputDTO);
-        var category = getCategory(shoppingListInputDTO.getCategoryId());
+        var category = categoryService.getById(shoppingListInputDTO.getCategoryId());
 
         if (shoppingListRepository.existsByNameAndAccountId(shoppingListInputDTO.getName(), accountId)) {
-            var errorMessage = String.format("Shopping list name already exists: %s", shoppingListInputDTO.getName());
-            log.error(errorMessage);
-            throw new ShoppingListException(HttpStatus.CONFLICT, errorMessage);
+            throw new ShoppingListException(HttpStatus.CONFLICT,
+                String.format("Shopping list name already exists: %s", shoppingListInputDTO.getName()));
         }
 
         shoppingList.setAccountId(accountId);
@@ -104,18 +95,8 @@ public class ShoppingListServiceImpl implements ShoppingListService {
         var savedShoppingList = shoppingListRepository.save(shoppingList);
 
         if (shoppingListInputDTO.getShared()) {
-            log.info("Setting shared with users for the shopping list, account id: {}, name: {}",
-                    accountId, shoppingListInputDTO.getName());
-            if (shoppingListInputDTO.getSharedWithAccounts() == null
-                    || shoppingListInputDTO.getSharedWithAccounts().isEmpty()) {
-                var errorMessage = "'sharedWithAccounts' cannot be null or empty while 'shared' is set to true";
-                log.error(errorMessage);
-                throw new ShoppingListException(HttpStatus.BAD_REQUEST, errorMessage);
-            }
-            List<SharedShoppingList> sharedLists = createSharedLists(
-                    savedShoppingList,
-                    shoppingListInputDTO.getSharedWithAccounts()
-            );
+            List<SharedShoppingList> sharedLists = sharedShoppingListService
+                .createSharedShoppingLists(savedShoppingList, shoppingListInputDTO.getSharedWithAccounts());
             savedShoppingList.getSharedShoppingLists().addAll(sharedLists);
         }
 
@@ -124,32 +105,21 @@ public class ShoppingListServiceImpl implements ShoppingListService {
         );
         return ShoppingListMapper.toDTO(savedShoppingList);
     }
+
     @Transactional(transactionManager = "shoppingListTransactionManager")
     public ShoppingListDTO update(UUID id, ShoppingListUpdateDTO shoppingListUpdateDTO) {
         log.info("Updating the shopping list: {}", id);
-        var optional = shoppingListRepository.findById(id);
-        var shoppingList = optional
-                .orElseThrow(() -> {
-                    var errorMessage = String.format(LIST_NOT_FOUND, id);
-                    log.error(errorMessage);
-                    return new ShoppingListException(HttpStatus.NOT_FOUND, errorMessage);
-                });
+        var shoppingList = getById(id);
         shoppingList.setName(shoppingListUpdateDTO.getName());
         shoppingList.setActive(shoppingListUpdateDTO.getActive());
-        shoppingList.setCategory(getCategory(shoppingListUpdateDTO.getCategoryId()));
+        shoppingList.setCategory(categoryService.getById(shoppingListUpdateDTO.getCategoryId()));
         return ShoppingListMapper.toDTO(shoppingListRepository.save(shoppingList));
     }
 
     @Transactional(transactionManager = "shoppingListTransactionManager")
     public ShoppingListDTO updateByFields(UUID id, ShoppingListUpdateDTO shoppingListUpdateDTO) {
         log.info("Updating the shopping list by fields: {}", id);
-        var optional = shoppingListRepository.findById(id);
-        var shoppingList = optional
-                .orElseThrow(() -> {
-                    var errorMessage = String.format(LIST_NOT_FOUND, id);
-                    log.error(errorMessage);
-                    return new ShoppingListException(HttpStatus.NOT_FOUND, errorMessage);
-                });
+        var shoppingList = getById(id);
         if (Objects.nonNull(shoppingListUpdateDTO.getName())) {
             shoppingList.setName(shoppingListUpdateDTO.getName());
         }
@@ -157,7 +127,7 @@ public class ShoppingListServiceImpl implements ShoppingListService {
             shoppingList.setActive(shoppingListUpdateDTO.getActive());
         }
         if (Objects.nonNull(shoppingListUpdateDTO.getCategoryId())) {
-            var category = getCategory(shoppingListUpdateDTO.getCategoryId());
+            var category = categoryService.getById(shoppingListUpdateDTO.getCategoryId());
             shoppingList.setCategory(category);
         }
 
@@ -166,41 +136,25 @@ public class ShoppingListServiceImpl implements ShoppingListService {
 
     public void delete(UUID id) {
         log.info("Deleting the shopping list: {}", id);
-        var shoppingList = shoppingListRepository.findById(id)
-                .orElseThrow(() -> {
-                    var errorMessage = String.format(LIST_NOT_FOUND, id);
-                    log.error(errorMessage);
-                    return new ShoppingListException(HttpStatus.NOT_FOUND, errorMessage);
-                });
+        final var shoppingList = shoppingListRepository.findById(id)
+            .orElseThrow(() -> new ShoppingListException(String.format("Shopping list not found: %s", id)));
+
+        var accountId = currentAccountService.getCurrentAccountId();
+        validateAccess(accountId, shoppingList);
         shoppingListRepository.delete(shoppingList);
     }
 
-    private Category getCategory(UUID categoryId) {
-        log.debug("Getting the shopping list category: {}", categoryId);
-        return categoryRepository.findById(categoryId)
-                .orElseThrow(() -> {
-                    var errorMessage = String.format("Category not found: %s", categoryId);
-                    log.error(errorMessage);
-                    return new CategoryException(HttpStatus.NOT_FOUND, errorMessage);
-                });
+    public Boolean validateAccess(UUID accountId, ShoppingList shoppingList) {
+        log.debug("Validating the shopping list access: {}", shoppingList.getId());
+
+        return !accountId.equals(shoppingList.getAccountId());
     }
 
-    private List<SharedShoppingList> createSharedLists(ShoppingList shoppingList, List<AccountSharedWithDto> sharedWith) {
-        log.debug("Creating shared lists for: {}", sharedWith);
-        return sharedWith.stream()
-                .map(accountSharedWithDto -> {
-                    final var email = accountSharedWithDto.getEmail();
-                    final var uuid = getAccountId(email);
-                    SharedShoppingList shared = new SharedShoppingList();
-                    shared.setShoppingListAndAccount(shoppingList, uuid);
-                    shared.setCostFactor(accountSharedWithDto.getCostFactor());
-                    return sharedShoppingListRepository.save(shared);
-                }).toList();
-    }
+    public Boolean validateSharedAccess(UUID accountId, ShoppingList shoppingList) {
+        log.debug("Validating the shopping list shared access");
 
-    private UUID getAccountId(String email) {
-        log.debug("Getting an account id for the email: {}", email);
-        return accountAPI.getAccountIdByEmail(email);
+        var accountIds = sharedShoppingListService.getAllAccountsByShoppingListId(shoppingList.getId());
+        return accountIds.stream().anyMatch(id -> id.equals(accountId));
     }
 
     private int calculateCostFactor(boolean isShared, List<AccountSharedWithDto> sharedWithAccounts) {
